+++ aworld/requirements.txt
@@ -1,6 +1,6 @@
 ######### [framework] #########
 # framework requirements will be installed by all extensions
-pydantic~=2.9.2
+pydantic>=2.9.2
 pyyaml~=6.0.2
 openai~=1.66.3
 mcp[cli]~=1.6.0
